refactor: move connection related entities together
