+++ src/connection/ConnectionRecoveryManager.ts
@@ -1,12 +1,12 @@
-import { WithSubscriptions } from './utils/WithSubscriptions';
-import { chatLoggerSystem } from './logger';
-import { runDetached } from './utils';
-import { ConfigController } from './configuration/ConfigController';
-import { deepFreezeConfig } from './configuration/utils/deepFreezeConfig';
-import type { StreamChat } from './client';
-import type { Channel } from './channel';
-import type { Thread } from './thread';
-import type { StateStore, Unsubscribe } from './store';
+import { WithSubscriptions } from '../utils/WithSubscriptions';
+import { chatLoggerSystem } from '../logger';
+import { runDetached } from '../utils';
+import { ConfigController } from '../configuration/ConfigController';
+import { deepFreezeConfig } from '../configuration/utils/deepFreezeConfig';
+import type { StreamChat } from '../client';
+import type { Channel } from '../channel';
+import type { Thread } from '../thread';
+import type { StateStore, Unsubscribe } from '../store';
 
 const logger = chatLoggerSystem.getLogger('client');
