+++ src/connection/wsConnection/index.ts
@@ -1,3 +1,4 @@
 export * from './config';
 export * from './types';
+export * from './StableWSConnection';
 export * from './WSConnection';
